findnvfsvb

+++ ShoeStore.BL/Services/ShoeService.cs
@@ -1,91 +1,73 @@
 ﻿using ShoeStore.BL.Interfaces;
 using ShoeStore.DL.Interfaces;
 using ShoeStore.Models.models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ShoeStore.BL.Services
 {
     public class ShoeService : IShoeService
     {
-    private readonly IShoeRepository _shoeRepository;
+        private readonly IShoeRepository _shoeRepository;
 
-    public ShoeService(IShoeRepository shoeRepository)
+        public ShoeService(IShoeRepository shoeRepository)
         {
             _shoeRepository = shoeRepository;
         }
 
-     public List<Shoe> GetAllShoes()
-     {
-            return _shoeRepository.GetAllShoes();
-      }
-
-    public Shoe GetShoe(int Id)
+        public List<Shoe> GetAllShoes()
         {
-            if (Id <= 0) return new Shoe();
+            return _shoeRepository.GetAllShoes();
+        }
+
+        public Shoe GetShoe(int Id)
+        {
+            if (Id <= 0)
+                return null; // Consider throwing an ArgumentException instead
             return _shoeRepository.GetShoe(Id);
         }
-    public void AddShoe(Shoe shoe)
+
+        public void AddShoe(Shoe shoe)
         {
             _shoeRepository.AddShoe(shoe);
-        }
-        public void RemoveShoe(Shoe shoe)
-        {
-            _shoeRepository.RemoveShoe(shoe);
-        }
-
-    public List<Shoe> GetAllByBrandAfterReleaseDate
-            (int brandId, DateTime afterDate)
-        {
-            var result = _shoeRepository.GetAllByBrand(brandId);
-
-            return result
-                .Where(b => b.ReleaseDate >= afterDate)
-                .ToList();
         }
 
         public void RemoveShoe(int Id)
         {
-            throw new NotImplementedException();
+            var shoe = _shoeRepository.GetShoe(Id);
+            if (shoe != null)
+                _shoeRepository.RemoveShoe(shoe);
+            // You may want to throw an exception if the shoe doesn't exist
         }
 
-        public List<Shoe> GetAllByBrandsAfterReleaseDate(int brandId, DateTime afterDate)
+        public List<Shoe> GetAllByBrandAfterReleaseDate(int brandId, DateTime afterDate)
         {
-            throw new NotImplementedException();
+            var result = _shoeRepository.GetAllByBrand(brandId);
+            return result.Where(b => b.ReleaseDate >= afterDate).ToList();
         }
 
+        // Explicit interface implementations
         List<Shoe> IShoeService.GetAllShoes()
         {
-            throw new NotImplementedException();
+            return GetAllShoes();
         }
 
         Shoe IShoeService.GetShoe(int Id)
         {
-            throw new NotImplementedException();
+            return GetShoe(Id);
         }
 
         void IShoeService.AddShoe(Shoe shoe)
         {
-            throw new NotImplementedException();
+            AddShoe(shoe);
         }
 
         void IShoeService.RemoveShoe(int Id)
         {
-            throw new NotImplementedException();
+            RemoveShoe(Id);
         }
 
         List<Shoe> IShoeService.GetAllByBrandsAfterReleaseDate(int brandId, DateTime afterDate)
         {
-            throw new NotImplementedException();
-        }
-
-        Shoe IShoeService.GetShoe()
-        {
-            throw new NotImplementedException();
+            return GetAllByBrandAfterReleaseDate(brandId, afterDate);
         }
     }
-
 }
